search function updated
