+++ src/components/ApplicationViews.js
@@ -8,7 +8,8 @@ import { CategoryProvider } from "./categories/CategoryProvider";
 
 import { DocumentProvider } from "./documents/DocumentProvider";
 import { DocumentForm } from "./documents/DocumentForm";
-import { DocumentSearch} from "./documents/DocumentSearch"
+import { DocumentSearch} from "./search/DocumentSearch"
+import { DocumentSearchList } from "./search/DocumentSearchList";
 
 import { SituationProvider } from "./situations/SituationProvider";
 
@@ -24,7 +25,7 @@ export const ApplicationViews = () => {
                         <Route path="documents" element={<Home />} />
                         <Route path="documents/create/*" element={<DocumentForm />} />
                         <Route path="documents/edit/:documentId/*" element={< DocumentForm />}/>
-                        <Route path="search" element={<DocumentSearch />} />
+                        <Route path="search" element={<><DocumentSearch /><DocumentSearchList /></>} />
                     </Routes>
                 </CustomerProvider>
             </CategoryProvider>
